+++ frontend/src/component/InPlaceEditableInput.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script>
-import api from '../api'
+import api from '@/api'
 
 export default {
   props: {
