+++ frontend/src/views/Itinerary/ItineraryView.vue
@@ -13,50 +13,12 @@
 
         <Notes :value="itinerary.notes" :itineraryId="itineraryId" />
         <Scratchpad :value="itinerary.scratchpad" :itineraryId="itineraryId" />
-
-        <Container>
-          <h4>Day-wise plan</h4>
-
-          <span v-for="day in itinerary.days" :key="day.id">
-            <Container>
-              <h3>
-                Day {{ day.order + 1 }}:
-                <InPlaceEditableInput
-                  inputType="input"
-                  placeholder="Write a title for you day"
-                  itemKey="name"
-                  :value="day.name"
-                  :editEndPoint="`/day/api/${day.id}/?itinerary_id=${itineraryId}`"
-                />
-              </h3>
-              <p>
-                <InPlaceEditableInput
-                  :value="day.notes"
-                  inputType="textarea"
-                  :editEndPoint="`/day/api/${day.id}/?itinerary_id=${itineraryId}`"
-                  itemKey="notes"
-                />
-              </p>
-              <button class="transparent-btn cross" @click="removeDay(day.id)">
-                X
-              </button>
-              <button
-                class="transparent-btn move-up"
-                @click="moveDay(day.id, 'up')"
-              >
-                ↑
-              </button>
-              <button
-                class="transparent-btn move-down"
-                @click="moveDay(day.id, 'down')"
-              >
-                ↓
-              </button>
-            </Container>
-            <!-- Display other details such as dishes, accommodations, restaurants, etc. as needed -->
-          </span>
-          <button class="green-btn" @click="addDay">Add Day</button>
-        </Container>
+        <DayWisePlan
+          :itineraryId="itineraryId"
+          :days="itinerary.days"
+          @updateDays="updateDays"
+          @fetchItinerary="fetchItinerary"
+        />
       </Container>
     </span>
 
@@ -67,11 +29,12 @@
 </template>
 
 <script>
-import api from '../../api'
-import Container from '../../component/Container.vue'
-import InPlaceEditableInput from '../../component/InPlaceEditableInput.vue'
-import Notes from '../../component/Itinerary/Notes.vue'
-import Scratchpad from '../../component/Itinerary/Scratchpad.vue'
+import api from '@/api'
+import Container from '@/component/Container.vue'
+import InPlaceEditableInput from '@/component/InPlaceEditableInput.vue'
+import DayWisePlan from '@/component/Itinerary/DayWisePlan.vue'
+import Notes from '@/component/Itinerary/Notes.vue'
+import Scratchpad from '@/component/Itinerary/Scratchpad.vue'
 
 export default {
   components: {
@@ -79,6 +42,7 @@ export default {
     InPlaceEditableInput,
     Notes,
     Scratchpad,
+    DayWisePlan,
   },
   data() {
     return {
@@ -87,7 +51,6 @@ export default {
     }
   },
   mounted() {
-    // Fetch itinerary details when the component is mounted
     this.fetchItinerary()
   },
   methods: {
@@ -102,68 +65,9 @@ export default {
           console.error('Error fetching itinerary:', error)
         })
     },
-    async addDay() {
-      const newDayData = {
-        name: 'New Day',
-        notes: '',
-      }
-      try {
-        const response = await api.post(
-          `/day/api/?itinerary_id=${this.itineraryId}`,
-          newDayData,
-        )
-        this.itinerary.days = [...this.itinerary.days, response.data]
-      } catch (error) {
-        // Handle errors
-        console.error('Error adding day:', error)
-      }
-    },
-    async removeDay(dayId) {
-      try {
-        await api.delete(`day/api/${dayId}/?itinerary_id=${this.itineraryId}`)
-        console.log('Day removed successfully:')
-        this.fetchItinerary()
-      } catch (error) {
-        console.error('Error removing day:', error)
-      }
-    },
-    async moveDay(dayId, moveDirection) {
-      try {
-        await api.post(
-          `day/${this.itineraryId}/${dayId}/move/${moveDirection}/`,
-        )
-        console.log('Day moved successfully:')
-        this.fetchItinerary()
-      } catch (error) {
-        console.error('Error moving day:', error)
-      }
+    updateDays(newValue) {
+      this.itinerary.days = newValue
     },
   },
 }
 </script>
-
-<style scoped>
-.cross {
-  color: red;
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  z-index: 1;
-}
-
-.move-up {
-  color: white;
-  position: absolute;
-  top: 30px;
-  right: 10px;
-  z-index: 1;
-}
-
-.move-down {
-  color: white;
-  position: absolute;
-  top: 50px;
-  right: 10px;
-  z-index: 1;
-}
-</style>
